- Migration to use plural format of test suffixes.
