+++ core/src/test/scala/za/co/absa/fadb/DBSchemaUnitTests.scala
@@ -19,7 +19,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import za.co.absa.fadb.naming.NamingConvention
 import za.co.absa.fadb.naming.implementations.SnakeCaseNaming.Implicits.namingConvention
 
-class DBSchemaUnitTest extends AnyFunSuite {
+class DBSchemaUnitTests extends AnyFunSuite {
 
   test("schema name default") {
     class Foo extends DBSchema
