+++ core/src/test/scala/za/co/absa/fadb/status/StatusExceptionUnitTests.scala
@@ -19,7 +19,7 @@ package za.co.absa.fadb.status
 import org.scalatest.funsuite.AnyFunSuite
 import za.co.absa.fadb.exceptions._
 
-class StatusExceptionUnitTest extends AnyFunSuite {
+class StatusExceptionUnitTests extends AnyFunSuite {
   test("Test equals - when they are the same") {
     val statusException = DataConflictException(FunctionStatus(10, "OK"))
     val otherStatusException = DataConflictException(FunctionStatus(10, "OK"))
