+++ doobie/src/test/scala/za/co/absa/fadb/doobie/DoobieMultipleResultFunctionIntegrationTests.scala
@@ -26,7 +26,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import za.co.absa.fadb.DBSchema
 import za.co.absa.fadb.doobie.DoobieFunction.DoobieMultipleResultFunction
 
-class DoobieMultipleResultFunctionIntegrationTest extends AnyFunSuite with DoobieTest {
+class DoobieMultipleResultFunctionIntegrationTests extends AnyFunSuite with DoobieTest {
 
   implicit def toFragmentsFunctionSemigroup[T]: Semigroup[T => Seq[Fragment]] = {
     (f1: T => Seq[Fragment], f2: T => Seq[Fragment]) => (params: T) => f1(params) ++ f2(params)
