+++ slick/src/test/scala/za/co/absa/fadb/slick/SlickMultipleResultFunctionIntegrationTests.scala
@@ -21,28 +21,29 @@ import org.scalatest.funsuite.AnyFunSuite
 import slick.jdbc.SQLActionBuilder
 import za.co.absa.fadb.DBSchema
 import za.co.absa.fadb.slick.FaDbPostgresProfile.api._
-import za.co.absa.fadb.slick.SlickFunction.SlickOptionalResultFunction
+import za.co.absa.fadb.slick.SlickFunction.SlickMultipleResultFunction
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class SlickOptionalResultFunctionIntegrationTest extends AnyFunSuite with SlickTest with ScalaFutures {
 
-  class GetActorById(implicit override val schema: DBSchema, val dbEngine: SlickPgEngine)
-      extends SlickOptionalResultFunction[Int, Actor]
+class SlickMultipleResultFunctionIntegrationTests extends AnyFunSuite with SlickTest with ScalaFutures {
+
+  class GetActors(implicit override val schema: DBSchema, val dbEngine: SlickPgEngine)
+    extends SlickMultipleResultFunction[GetActorsQueryParameters, Actor]
       with ActorSlickConverter {
 
     override def fieldsToSelect: Seq[String] = super.fieldsToSelect ++ Seq("actor_id", "first_name", "last_name")
 
-    override protected def sql(values: Int): SQLActionBuilder = {
-      sql"""SELECT #$selectEntry FROM #$functionName($values) #$alias;"""
+    override def sql(values: GetActorsQueryParameters): SQLActionBuilder = {
+      sql"""SELECT #$selectEntry FROM #$functionName(${values.firstName},${values.lastName}) #$alias;"""
     }
   }
 
-  private val getActorById = new GetActorById()(Integration, new SlickPgEngine(db))
+  private val getActors = new GetActors()(Integration, new SlickPgEngine(db))
 
-  test("Retrieving an actor by id from database") {
-    val expectedResultElem = Some(Actor(49, "Pavel", "Marek"))
-    val results = getActorById(49)
-    assert(results.futureValue == expectedResultElem)
+  test("Retrieving actors from database") {
+    val expectedResultElem = Actor(49, "Pavel", "Marek")
+    val results = getActors(GetActorsQueryParameters(Some("Pavel"), Some("Marek")))
+    assert(results.futureValue.contains(expectedResultElem))
   }
 }
